Adjust vim terminal size
The default terminal size is half the screen, which is too large
for the shellcheck output. The terminal is scrollable, so 12 rows
should be plenty to see a few errors at a time.

Also my 'Save and Exit' binding was giving me trouble with the terminal
since it becomes a read-only buffer when the command completes. This
means you can't save it, and the only way to close it is to do ':q'.
To make the '<Leader>q' binding work, I need to remove the save part,
which is not a problem since I usually save before exiting anyway.